Avancement sur le graphics_updater

+++ src/client.c
@@ -5,12 +5,11 @@
 #include <unistd.h>
 #include <sys/wait.h>
 
-#include "level.h"
 #include "utils.h"
 #include "constants.h"
 
-#include "client_gui.h"
 #include "client_network.h"
+#include "gui_updater.h"
 
 // https://gitlab-mi.univ-reims.fr/rabat01/info0601/-/blob/main/Cours/01_curses/CM_01.pdf
 
@@ -22,20 +21,6 @@ int pid_network, pid_updater_graphics;
 void stop_game() {
 	// Logs
 	logs(L_INFO, "Main | Stopping game...");
-
-	// Free the game interface
-    stop_gui();
-}
-
-/**
- * @brief Function managing the keyboard events.
- * Running indefinitely until the user press the key to quit the game
- */
-void control_handler() {
-    int ch;
-    while((ch = getch()) != KEY_QUIT_GAME) {
-        control_handler_gui(ch);
-    }
 }
 
 /**
@@ -43,11 +28,22 @@ void control_handler() {
  */
 void main_exit() {
     stop_game();
-    kill(pid_network, SIGINT);
-    waitpid(pid_network, NULL, 0);
 
-    kill(pid_updater_graphics, SIGINT);
-    waitpid(pid_updater_graphics, NULL, 0);
+    // check if the network is still running
+    logs(L_INFO, "Main | Check network...");
+    if (kill(pid_network, 0) == 0) {
+        logs(L_INFO, "Main | Stopping network...");
+        kill(pid_network, SIGINT);
+        waitpid(pid_network, NULL, 0);
+    }
+
+    // check if the updater is still running
+    logs(L_INFO, "Main | Check updater...");
+    if (kill(pid_updater_graphics, 0) == 0) {
+        logs(L_INFO, "Main | Stopping updater...");
+        kill(pid_updater_graphics, SIGINT);
+        waitpid(pid_updater_graphics, NULL, 0);
+    }
 
     // Close logs
     closeLogs();
@@ -64,11 +60,11 @@ int main(int argc, char *argv[]) {
     // Register exit function
     atexit(main_exit);
 
-    // Graphics updater
+    // Graphics updater & control handler
     pid_updater_graphics = init_updater_gui();
 
-    // Launch control handler
-    control_handler();
+    // Wait for the end of the game
+    waitpid(pid_updater_graphics, NULL, 0);
 
 	return EXIT_SUCCESS;
 }
